refactor: 전체 회고 개수 조회 SQL 문 개선  - 단순 전체 개수를 조회하는데 전체 컬럼이 select 되고 있어 수정

+++ dailyge-api/src/main/java/project/dailyge/app/core/retrospect/application/RetrospectReadService.java
@@ -10,5 +10,5 @@ public interface RetrospectReadService {
 
     List<RetrospectJpaEntity> findRetrospectByPage(DailygeUser dailygeUser, CustomPageable page);
 
-    int findTotalCount(DailygeUser dailygeUser);
+    long findTotalCount(DailygeUser dailygeUser);
 }
